move pd scheduler to separate struct
