+++ pkg/task/backup.go
@@ -217,7 +217,7 @@ func RunBackup(c context.Context, g glue.Glue, cmdName string, cfg *BackupConfig
 
 	if cfg.RemoveSchedulers {
 		log.Debug("removing some PD schedulers")
-		restore, e := mgr.RemoveSchedulers(ctx)
+		restore, e := mgr.PdMgr.RemoveSchedulers(ctx)
 		defer func() {
 			if restoreE := restore(ctx); restoreE != nil {
 				log.Warn("failed to restore removed schedulers, you may need to restore them manually", zap.Error(restoreE))
