+++ pkg/task/backup_raw.go
@@ -142,7 +142,7 @@ func RunBackupRaw(c context.Context, g glue.Glue, cmdName string, cfg *RawKvConf
 	backupRange := rtree.Range{StartKey: cfg.StartKey, EndKey: cfg.EndKey}
 
 	if cfg.RemoveSchedulers {
-		restore, e := mgr.RemoveSchedulers(ctx)
+		restore, e := mgr.PdMgr.RemoveSchedulers(ctx)
 		defer func() {
 			if restoreE := restore(ctx); restoreE != nil {
 				log.Warn("failed to restore removed schedulers, you may need to restore them manually", zap.Error(restoreE))
